requirements.txt
paho.mqtt
pyyaml

